fix(audit): pass 4 — split navigation_actions under 800 lines + strengthen adjacency test
- navigation_actions.cpp was 801 lines, 1 over the CLAUDE.md <800 rule. Extract
  the four cross-surface resolution helpers (resolveCrossDesktopZone,
  tryCrossModeOutput, windowInZoneOnScreen, entryZoneForCrossing) into a new
  navigation_crosssurface.cpp (they use no file-local helpers and are called by
  the staying code only via member calls). navigation_actions.cpp → 716 lines.
- test_spatial_adjacency skipsCurrentByCentre: the old form (current + one
  neighbour) didn't discriminate the centre-skip from plain direction filtering.
  Add a co-centred duplicate between current and the neighbour so the skip path
  is actually exercised and a regression admitting equal-centre candidates fails.
