+++ libs/phosphor-snap-engine/src/navigation_actions.cpp
@@ -397,91 +397,6 @@ bool SnapEngine::tryCrossDesktopMove(const QString& windowId, const QString& dir
     return true;
 }
 
-std::pair<QString, QRect> SnapEngine::resolveCrossDesktopZone(const QString& currentZoneId, const QString& screenId,
-                                                              int targetDesktop) const
-{
-    PhosphorZones::Layout* targetLayout =
-        m_layoutManager ? m_layoutManager->layoutForScreen(screenId, targetDesktop, currentActivity()) : nullptr;
-    if (!targetLayout || targetLayout->zones().isEmpty()) {
-        return {};
-    }
-    // Position is the 1-based enumeration index of zones sorted by zone number
-    // (NOT the raw zone number) — the same scheme the resnap path uses.
-    const int position =
-        PhosphorZones::LayoutUtils::buildGlobalZonePositionMap(m_layoutManager->layouts()).value(currentZoneId, 0);
-    QVector<PhosphorZones::Zone*> zones = targetLayout->zones();
-    PhosphorZones::LayoutUtils::sortZonesByNumber(zones);
-    PhosphorZones::Zone* targetZone =
-        (position >= 1 && position <= zones.size()) ? zones.value(position - 1, nullptr) : nullptr;
-    if (!targetZone) {
-        return {};
-    }
-    const QString targetZoneId = targetZone->id().toString();
-    const QRect geo = m_windowTracker ? m_windowTracker->zoneGeometry(targetZoneId, screenId) : QRect();
-    if (!geo.isValid()) {
-        return {};
-    }
-    return {targetZoneId, geo};
-}
-
-bool SnapEngine::tryCrossModeOutput(const QString& windowId, const QString& direction, const QString& screenId,
-                                    bool swap)
-{
-    if (!m_crossSurfaceResolver || !m_layoutManager) {
-        return false;
-    }
-    const QString neighbour = m_crossSurfaceResolver->neighborOutputInDirection(screenId, direction);
-    if (neighbour.isEmpty()) {
-        return false;
-    }
-    // Reaching here means the resolver found no snap entry zone on the neighbour
-    // (the resolver returned no_adjacent_zone). Only an AUTOTILE neighbour is a
-    // cross-mode handoff; a snap neighbour with no entry zone is a genuine
-    // boundary — leave it. A move inserts the window into the neighbour's stack;
-    // a swap trades it with the neighbour's entry-edge tile.
-    if (m_layoutManager->modeForScreen(neighbour, currentVirtualDesktop(), currentActivity())
-        != PhosphorZones::AssignmentEntry::Autotile) {
-        return false;
-    }
-    if (swap) {
-        Q_EMIT crossModeSwapRequested(windowId, neighbour, 0, direction);
-    } else {
-        Q_EMIT crossModeMoveRequested(windowId, neighbour, 0, direction);
-    }
-    return true;
-}
-
-QString SnapEngine::windowInZoneOnScreen(const QString& zoneId, const QString& screenId) const
-{
-    if (!m_windowTracker || zoneId.isEmpty()) {
-        return QString();
-    }
-    // windowsInZone is screen-agnostic (a zone UUID is shared by every output the
-    // layout drives), so pin to the daemon's stored screen assignment.
-    const QStringList windows = m_windowTracker->windowsInZone(zoneId);
-    const QHash<QString, QString>& screens = m_windowTracker->screenAssignments();
-    for (const QString& windowId : windows) {
-        if (screens.value(windowId) == screenId) {
-            return windowId;
-        }
-    }
-    return QString();
-}
-
-QString SnapEngine::entryZoneForCrossing(const QString& direction, const QString& neighbourScreen) const
-{
-    if (!m_zoneAdjacencyResolver) {
-        return {};
-    }
-    // Enter the neighbour from the edge facing back toward the source (crossing
-    // "right" lands on the neighbour's LEFT-edge zone, etc.) — shared mapping.
-    const QString opposite = oppositeCrossingDirection(direction);
-    if (opposite.isEmpty()) {
-        return {};
-    }
-    return m_zoneAdjacencyResolver->getFirstZoneInDirection(opposite, neighbourScreen);
-}
-
 void SnapEngine::swapFocusedInDirection(const QString& direction, const NavigationContext& ctx)
 {
     qCInfo(PhosphorSnapEngine::lcSnapEngine) << "SnapEngine::swapFocusedInDirection:" << direction;
